adjust datatype of count and time, add mode template

+++ Algorithms/BubbleSort.cpp
@@ -5,7 +5,7 @@
 
 class BubbleSort : public Sort {
 protected:
-    void Handle(int &count_compare) {
+    void Handle(ull &count_compare) {
         for (int  i = 0; ++count_compare && i < size - 1; i++)
         {
        		for (int j = 0; ++count_compare && j < size - 1 - i; j++)
